refactor(json): consolidate guava codecs
